+++ src/components/requestDetail/RequestContent.vue
@@ -7,7 +7,7 @@ import { useUserStore } from '/@/stores/user'
 import InputTextarea from '/@/components/shared/InputTextarea.vue'
 import MarkdownIt from '/@/components/shared/MarkdownIt.vue'
 import SimpleButton from '/@/components/shared/SimpleButton.vue'
-import type { EditMode } from '/@/pages/composables/requestDetail/useRequestDetail'
+import type { EditMode } from '/@/pages/composables/useRequestDetail'
 
 interface Props {
   isEditMode: boolean
