+++ src/components/requestDetail/RequestTitle.vue
@@ -6,7 +6,7 @@ import { useUserStore } from '/@/stores/user'
 
 import InputText from '/@/components/shared/InputText.vue'
 import SimpleButton from '/@/components/shared/SimpleButton.vue'
-import type { EditMode } from '/@/pages/composables/requestDetail/useRequestDetail'
+import type { EditMode } from '/@/pages/composables/useRequestDetail'
 
 interface Props {
   isEditMode: boolean
